requirements.txt
pandas~=1.3.5
plotly~=5.6.0
flask~=2.0.3
Gunicorn
psycopg2~=2.9.3
requests~=2.27.1
sklearn~=0.0
scipy~=1.8.0
SQLAlchemy~=1.4.32
pip~=21.3.1
wheel~=0.37.1
cryptography~=36.0.2
lxml~=4.8.0
tornado~=6.1
pytz~=2021.3
MarkupSafe~=2.0.1
Werkzeug~=2.0.3
Jinja2~=3.0.3
click~=8.0.3
itsdangerous~=2.1.1
numpy~=1.22.1
setuptools~=60.5.0
cffi~=1.15.0
matplotlib~=3.5.1
joblib~=1.1.0
python-dateutil~=2.8.2
six~=1.16.0
tenacity~=8.0.1
ipywidgets~=7.7.0
traitlets~=5.1.1
certifi~=2021.10.8
scikit-learn~=1.0.2
threadpoolctl~=3.1.0
Pillow~=9.0.0
greenlet~=1.1.2
idna~=3.3
urllib3~=1.26.8
packaging~=21.3
attrs~=21.4.0